Fix: 小bug修復 / 整理程式碼 / html模板 / 後端MVC / Uyils regex  12/20
