+++ flask/api/attractions/attractions.py
@@ -1,4 +1,4 @@
-from flask import Flask,Blueprint,jsonify,request,render_template
+from flask import Blueprint,jsonify,request,render_template
 from api.models.attractions_model import Attraction
 
 
@@ -24,35 +24,12 @@ def get_all_attractions():
         # 如果沒有輸入page參數 或 如果page小於0 或 如果page不是數字
         if page == None or page.isdigit() != True or int(page)<0:
             page_data["nextPage"] = None
-            #page_data["data"].append("參數錯誤，無法搜尋資料")
             return jsonify(page_data),400
-        page = int(page)+1
         # 有無keyword
         if keyword == None:
-            pages = Attraction.get_attraction_pages(page)
+            page_data = Attraction.get_attraction_pages(page)
         else:
-            pages = Attraction.get_attraction_pages_by_keywords(keyword,page)
-        # 頁數判斷，如果下一頁沒有資料就顯示null
-        if page >= pages.pages:
-            page_data["nextPage"] = None
-        # 如果沒資料，不用再寫判斷式，data會自動留空
-        for data in pages:
-            image_urls = []
-            for image in data.images:
-                image_urls.append(image.image_url)
-            attraction = {
-                "id" : data.id,
-                "name" : data.name,
-                "category" : data.category,
-                "description" : data.description,
-                "address" : data.address,
-                "transport" : data.transport,
-                "mrt" : data.mrt,
-                "lat" : data.lat,
-                "lng" : data.lng,
-                "images" : image_urls
-            }
-            page_data["data"].append(attraction)
+            page_data = Attraction.get_attraction_pages_by_keywords(keyword,page)
         return jsonify(page_data),200
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
@@ -61,28 +38,10 @@ def get_all_attractions():
 @attractions.route("api/attraction/<attractionId>",methods=["GET"])
 def get_attraction(attractionId):
     try:
-        query = Attraction.get_attraction_data_by_attractionId(attractionId)
-        # 錯誤可能
-        if query == None:
+        attraction_data = Attraction.get_attraction_data_by_attractionId(attractionId)
+        if attraction_data == "no_data":
             return jsonify(error="true",message="沒有此景點編號"),400
-        image_urls = []
-        for image in query.images:
-            image_urls.append(image.image_url)
-        attraction = {
-            "data": {
-                "id" :query.id,
-                "name" : query.name,
-                "category" : query.category,
-                "description" : query.description,
-                "address" : query.address,
-                "transport" :query.transport,
-                "mrt" : query.mrt,
-                "lat" : query.lat,
-                "lng" : query.lng,
-                "images" : image_urls
-            }
-        }
-        return jsonify(attraction),200
+        return jsonify(attraction_data),200
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
         
@@ -91,14 +50,7 @@ def get_attraction(attractionId):
 @attractions.route("api/categories",methods=["GET"])
 def get_attraction_categories():
     try:
-        # 不重複查詢
-        query = Attraction.get_attraction_categories()
-        categories = []
-        for data in query:
-           categories.append(data[0])
-        category_data = {
-            "data" : categories
-        }
+        category_data = Attraction.get_attraction_categories()
         return jsonify(category_data),200
     except Exception as ex:
         return jsonify(error="true",message=f"{ex}"),500
