+++ flask/api/static/js/fetchAPI.js
@@ -1,5 +1,37 @@
+// fetch 景點資料
+export async function fetchAttractionPageData(page=0,keyword=""){
+    try{
+        const response = await fetch(`/api/attractions?page=${page}&keyword=${keyword}`);
+        let data = await response.json()
+        return data
+    }
+    catch(err){
+        console.log("fetch failed:",err)
+    }
+}
 
-import { refreshAccessToken } from "./member.js"
+export async function fetchCategoryData(page=0,keyword=""){
+    try{
+        const response = await fetch("/api/categories");
+        let data = await response.json()
+        return data
+    }
+    catch(err){
+        console.log("fetch failed:",err)
+    }
+}
+
+
+export async function fetchAttraction(id=1){
+    try{
+        const response = await fetch(`/api/attraction/${id}`);
+        let data = await response.json()
+        return data
+    }
+    catch(err){
+        console.log("fetch failed:",err)
+    }
+}
 
 /* register */
 export async function register(name,email,password){
