--- flask/api/static/js/fetchLocation.js
@@ -1,38 +1,0 @@
-// fetch 景點資料
-
-
-export async function fetchAttractionPageData(page=0,keyword=""){
-    try{
-        const response = await fetch(`/api/attractions?page=${page}&keyword=${keyword}`);
-        let data = await response.json()
-        return data
-    }
-    catch(err){
-        console.log("fetch failed:",err)
-    }
-}
-
-export async function fetchCategoryData(page=0,keyword=""){
-    try{
-        const response = await fetch("/api/categories");
-        let data = await response.json()
-        return data
-    }
-    catch(err){
-        console.log("fetch failed:",err)
-    }
-}
-
-
-export async function fetchAttraction(id=1){
-    try{
-        const response = await fetch(`/api/attraction/${id}`);
-        let data = await response.json()
-        return data
-    }
-    catch(err){
-        console.log("fetch failed:",err)
-    }
-}
-
-
